# Improve Union deserialization performance