Use `@generated` to ensure that the deserialization of `Union`s is type stable.

Using these 4 structs as examples
```Julia
using JSON3, StructTypes, BenchmarkTools
struct T1
    x::Int
end
struct T2
    x::Union{Nothing, Int}
end
struct T3
    x::Int
end
struct T4
    x::Union{Nothing, Int}
end
StructTypes.StructType(::Type{T1}) = StructTypes.Struct()
StructTypes.StructType(::Type{T2}) = StructTypes.Struct()
StructTypes.StructType(::Type{T3}) = StructTypes.OrderedStruct()
StructTypes.StructType(::Type{T4}) = StructTypes.OrderedStruct()
const s  = "{\"x\":1}"
```

Before this PR:

```Julia
julia> @btime JSON3.read($s, T1)
  142.779 ns (1 allocation: 336 bytes)
julia> @btime JSON3.read($s, T2)
  738.870 ns (7 allocations: 512 bytes)
julia> @btime JSON3.read($s, T3)
  29.438 ns (0 allocations: 0 bytes)
julia> @btime JSON3.read($s, T4)
  787.804 ns (7 allocations: 192 bytes)
```

After this PR:

```Julia
julia> @btime JSON3.read($s, T1)
  143.453 ns (1 allocation: 336 bytes)
julia> @btime JSON3.read($s, T2)
  290.414 ns (2 allocations: 384 bytes)
julia> @btime JSON3.read($s, T3)
  29.438 ns (0 allocations: 0 bytes)
julia> @btime JSON3.read($s, T4)
  34.407 ns (0 allocations: 0 bytes)
```

The additional allocation when deserialising the `UnorderedStruct` with the union occurs (I think) in the `StructClosure`, and I'm not sure if this can be removed.

Ref #152