@mmiller-max 2021-09-28
Also a benchmark without `Nothing`

```Julia
struct T5
       val::Union{Float64, String}
end
StructTypes.StructType(::Type{T5}) = StructTypes.OrderedStruct()
str1 = "{\"val\": \"str\"}"
str2 = "{\"val\": 0.5}"
```
Before PR:
```Julia
julia> @btime JSON3.read($str1, T5)
  59.500 μs (29 allocations: 1.11 KiB)
T5("str")

julia> @btime JSON3.read($str2, T5)
  1.000 μs (13 allocations: 304 bytes)
T5(0.5)
```
After PR:
```Julia
julia> @btime JSON3.read($str1, T5)
  49.833 μs (20 allocations: 928 bytes)
T5("str")

julia> @btime JSON3.read($str2, T5)
  196.294 ns (2 allocations: 48 bytes)
T5(0.5)
```